Make CMS menu editable

+++ src/KikCMS/Controllers/FinderController.php
@@ -98,6 +98,10 @@ class FinderController extends RenderableController
 
         $filePath = $this->finderFileService->getFilePath($finderFile);
 
+        if( ! file_exists($filePath)){
+            throw new NotFoundException();
+        }
+
         return $this->outputFile($filePath, $finderFile->getMimeType(), $finderFile->getName());
     }
